chore(java): separate classes into packages
- move car related classes to car package
- move game related classes to game package
