+++ src/main/java/racingcar/car/CarNames.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.car;
 
 import java.util.HashSet;
 import java.util.Set;
