+++ src/main/java/racingcar/car/RacingCars.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.car;
 
 import java.util.List;
 import java.util.Optional;
