+++ src/main/java/racingcar/game/GameRound.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.game;
 
 import view.OutputView;
